fix: add option to disable MPI (#68)
CalcJob plugins can specify default values for whether codes should use
MPI or not: using the `metadata.options.withmpi` variable.

aiida-testing can run calculations with a single MPI process if an MPI
library is installed. However, the MPI library is then also required in
order to fetch the results from the test database, because the mock
executable will be run with `mpirun -np 1 ...`.

Here we add an option `--mock-disable-mpi` that allows users to disable
the use of MPI. If supported by the AiiDA version (>=2.1.0), this option is
enabled automatically by `--mock-fail-on-missing`.
The documentation now recommends using `--mock-fail-on-missing` on CI
by default.