Support aiida<2.6 (#84)
mock_code tests were failing since aiida 2.3.0,
caused by just a whitespace change
in the core.direct scheduler submission script.

Concretely, there was an extra space at the end of the last
line of the script (the executable line) before 2.3.
As a short term solution to preserve the existing hashes,
we append a space if the script doesn't end with one.

This change is purely for backwards compatibility and should be
reverted right after 0.0.1 release.